requirements.txt
fastapi==0.70.0
databases==0.7.2
uvicorn==0.15.0
face-recognition==1.3.0
Pillow==8.4.0
numpy==1.21.2
pydantic==1.9.0
sqlalchemy==1.4.25
